requirements.txt
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Werkzeug==2.0.2
pymysql==1.1.1
